Update Core_Calc.py
Code has been shortened, below is a an explanation of what I have done. 

math.floor(28/7), math.floor(29/7), math.floor(30/7) and math.floor(31/7) all give an output of 4. This number 4 refers to how many weeks any month can have. The total amount of core from the weekly task is therefore 4 * 15 = 60. Thus there is no need for a week_core variable. 

Condensed a 4 line code or so with a ternary operator.